fix/#32: 점수 타입 Long으로 변경

+++ src/main/java/com/project/common/util/MessageFormatUtil.java
@@ -25,7 +25,7 @@ public class MessageFormatUtil {
     return sb.toString();
   }
 
-  public String formatRankChange(String userName, int oldRank, int newRank, int score) {
+  public String formatRankChange(String userName, int oldRank, int newRank, long score) {
     return String.format(
         """
                         🔥 %s님이 %d위 → %d위로 상승! 🔥
